requirements.txt
boto3
pandas
python-dotenv
setuptools
git+https://github.com/validityBase/vbase-py.git#egg=vbase
